Restructure agent and dockerify its tests.
