+++ agent/tox.ini
@@ -1,29 +1,78 @@
 [tox]
-envlist = py37
+envlist = docker
+
+[testenv:docker]
+whitelist_externals=docker
+passenv =
+    PYTHON_VERSION
+    TOXENV
+setenv =
+    PYTHON_VERSION=3.7
+    TOXENV=check,py37
+deps =
+usedevelop = false
+depends =
+skip_install = true
+commands =
+    docker build --build-arg PYTHON_VERSION={env:PYTHON_VERSION} -t balrog_agent-{envname}-test -f Dockerfile .
+    docker run --rm -v {toxinidir}:/app -v balrog_agent-{envname}-tox:/app/.tox -e TOXENV={env:TOXENV} --entrypoint /app/scripts/test-entrypoint.sh balrog_agent-{envname}-test
 
 [testenv]
-# ignore_errors makes sure all commands get run, which means this won't abort if flake8 fails.
-# it doesn't affect the return code in the end.
-ignore_errors = True
+depends = clean
 setenv =
     PYTHONDONTWRITEBYTECODE=1
-
-deps = -rrequirements/test.txt
-
+    HYPOTHESIS_STORAGE_DIRECTORY=/tmp
+usedevelop = True
+deps =
+    -r {toxinidir}/requirements/test.txt
 commands =
-    flake8 balrogagent
-    py.test --cov=balrogagent {posargs:balrogagent}
+    {posargs:py.test -n auto --cov-config=tox.ini --cov-append --cov=balrogagent --cov-report term-missing tests}
+
+[testenv:clean]
+skip_install = true
+commands = coverage erase
+depends =
+
+[testenv:report]
+skip_install = true
+commands = coverage report -m
+depends = py37
+parallel_show_output = true
+
+[testenv:check]
+skip_install = true
+commands =
+    black --check {toxinidir}
+    isort --check -rc -df {toxinidir}
+    flake8 {toxinidir}
+    pip-compile-multi verify
+# Doesn't work without a .git directory
+#    check-manifest -v {toxinidir}
 
 [flake8]
 max-line-length = 160
-exclude = .tox, .git, build
+exclude = .ropeproject,.tox,sandbox,build
 show-source = True
+# flake8 doesn't differentiat the use of `:` in dictionaries vs. slices,
+# whereas # black formats the a later with surrounding whitespace. This causes
+# flake8 to incorrecly report E203.
+# See https://black.readthedocs.io/en/stable/the_black_code_style.html#slices
+# W503 is a default ignore, but gets overridden when we set this, so we need
+# to add it again ourselves.
+ignore = E203,W503
 
 [pytest]
-norecursedirs = .tox .git build
+norecursedirs = .tox .git .hg sandbox build
 python_files = test_*.py
-filterwarnings =
-    ignore
-    error:::balrogagent
-    default::DeprecationWarning:balrogagent
-    default::PendingDeprecationWarning:balrogagent
+
+[coverage:run]
+branch = true
+source = balrogagent
+
+[coverage:paths]
+source =
+  src/
+  .tox/*/lib/python*/site-packages/
+
+[coverage:report]
+show_missing = True
