+++ setup.py
@@ -36,7 +36,6 @@ setup(
     include_package_data=True,
     install_requires=requirements,
     test_suite="tests",
-    tests_require="test_requirements",
     url="https://github.com/mozilla/balrog",
     license="MPL-2.0",
     zip_safe=False,
